Move embedding-SDK plugin into its own metabase/embedding/sdk directory
Make the plugin layout consistent with the moderation plugin: a plugin owns a
dedicated directory holding plugin.ts (the OSS default / build-override target)
and types.ts (the contract). This drops the .types filename suffix and stops
the plugin's files sitting loose in the shared metabase/embedding directory.

  metabase/embedding/sdk-plugin.ts        -> metabase/embedding/sdk/plugin.ts
  metabase/embedding/sdk-plugin.types.ts  -> metabase/embedding/sdk/types.ts

Consumers now import metabase/embedding/sdk/plugin; the rspack alias and the
contract import are updated to match. No module-boundary change is needed — the
files stay in the shared/embedding module.
