+++ frontend/src/metabase/admin/ai/AISettingsPage.tsx
@@ -12,7 +12,7 @@ import { Link } from "metabase/common/components/Link";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
 import { UpsellGem } from "metabase/common/components/upsells/components/UpsellGem";
 import { useDocsUrl, useSetting } from "metabase/common/hooks";
-import { PLUGIN_EMBEDDING_SDK } from "metabase/embedding/sdk-plugin";
+import { PLUGIN_EMBEDDING_SDK } from "metabase/embedding/sdk/plugin";
 import { FIXED_METABOT_IDS } from "metabase/metabot/constants";
 import { PLUGIN_EMBEDDING_IFRAME_SDK } from "metabase/plugins";
 import { useRouter } from "metabase/router/useRouter";
