+++ frontend/src/metabase/embedding/sdk/plugin.ts
@@ -1,6 +1,6 @@
 import type { OnBeforeRequestHandler } from "metabase/plugins/oss/api";
 
-import type { EmbeddingSdkPlugin } from "./sdk-plugin.types";
+import type { EmbeddingSdkPlugin } from "./types";
 
 const noopRequestHandler: OnBeforeRequestHandler = async () => {};
 
@@ -11,7 +11,7 @@ const noopRequestHandler: OnBeforeRequestHandler = async () => {};
  * `metabase-enterprise/embedding-sdk/plugin` (see resolve-aliases.js), so this
  * file is never bundled there. Consumers import the resolved module directly:
  *
- *   import { PLUGIN_EMBEDDING_SDK } from "metabase/embedding/sdk-plugin";
+ *   import { PLUGIN_EMBEDDING_SDK } from "metabase/embedding/sdk/plugin";
  *
  * `onBeforeRequestHandlers` is a runtime-mutable holder: writers reassign its
  * fields during SDK/guest-embed init and the API middleware reads them.
